+++ Content.Shared/GameObjects/Components/Chemistry/SharedSolutionComponent.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.Collections.Generic;
+using Content.Shared.Chemistry;
+using Robust.Shared.GameObjects;
+using Robust.Shared.IoC;
+using Robust.Shared.Maths;
+using Robust.Shared.Prototypes;
+using Robust.Shared.Serialization;
+using Robust.Shared.ViewVariables;
+
+namespace Content.Shared.GameObjects.Components.Chemistry
+{
+    public class SharedSolutionComponent : Component
+    {
+        public override string Name => "Solution";
+
+        /// <inheritdoc />
+        public sealed override uint? NetID => ContentNetIDs.SOLUTION;
+
+        [Serializable, NetSerializable]
+        public class SolutionComponentState : ComponentState
+        {
+            public SolutionComponentState() : base(ContentNetIDs.SOLUTION) { }
+        }
+
+        /// <inheritdoc />
+        public override ComponentState GetComponentState()
+        {
+            return new SolutionComponentState();
+        }
+
+        /// <inheritdoc />
+        public override void HandleComponentState(ComponentState curState, ComponentState nextState)
+        {
+            base.HandleComponentState(curState, nextState);
+
+            if(curState == null)
+                return;
+
+            var compState = (SolutionComponentState)curState;
+
+            //TODO: Make me work!
+        }
+
+    }
+}
