+++ Resources/Prototypes/Entities/buildings/chem_dispenser.yml
@@ -37,3 +37,4 @@
   - chem.K
   - chem.Ra
   - chem.Na
+  - chem.U
