+++ Resources/Prototypes/Entities/buildings/soda_dispenser.yml
@@ -19,3 +19,4 @@
   - chem.Tea
   - chem.Ice
   - chem.H2O
+  - chem.Cream
