+++ Resources/Prototypes/Entities/items/Consumables/drinks_cans.yml
@@ -30,11 +30,12 @@
   name: Space cola
   description: A refreshing beverage.
   components:
-  - type: Drink    
+  - type: Solution
+    maxVol: 20
     contents:
       reagents:
-      - ReagentId: chem.H2O
-        Quantity: 4
+      - ReagentId: chem.Cola
+        Quantity: 20
   - type: Sprite
     sprite: Objects/Drinks/cola.rsi
   - type: Icon
@@ -62,10 +63,6 @@
   description: ''
   components:
   - type: Drink    
-    contents:
-      reagents:
-      - ReagentId: chem.H2O
-        Quantity: 4
   - type: Sprite
     sprite: Objects/Drinks/ice_tea_can.rsi
   - type: Icon
@@ -93,10 +90,6 @@
   description: You wanted ORANGE. It gave you Lemon Lime.
   components:
   - type: Drink    
-    contents:
-      reagents:
-      - ReagentId: chem.H2O
-        Quantity: 4
   - type: Sprite
     sprite: Objects/Drinks/lemon-lime.rsi
   - type: Icon
@@ -124,10 +117,6 @@
   description: ''
   components:
   - type: Drink    
-    contents:
-      reagents:
-      - ReagentId: chem.H2O
-        Quantity: 4
   - type: Sprite
     sprite: Objects/Drinks/purple_can.rsi
   - type: Icon
@@ -155,10 +144,6 @@
   description: Blows right through you like a space wind.
   components:
   - type: Drink    
-    contents:
-      reagents:
-      - ReagentId: chem.H2O
-        Quantity: 4
   - type: Sprite
     sprite: Objects/Drinks/space_mountain_wind.rsi
   - type: Icon
@@ -186,10 +171,6 @@
   description: Tastes like a hull breach in your mouth.
   components:
   - type: Drink    
-    contents:
-      reagents:
-      - ReagentId: chem.H2O
-        Quantity: 4
   - type: Sprite
     sprite: Objects/Drinks/space-up.rsi
   - type: Icon
@@ -217,10 +198,6 @@
   description: The taste of a star in liquid form. And, a bit of tuna...?
   components:
   - type: Drink    
-    contents:
-      reagents:
-      - ReagentId: chem.H2O
-        Quantity: 4
   - type: Sprite
     sprite: Objects/Drinks/starkist.rsi
   - type: Icon
@@ -248,10 +225,6 @@
   description: The MBO has advised crew members that consumption of Thirteen Loko may result in seizures, blindness, drunkeness, or even death. Please Drink Responsibly.
   components:
   - type: Drink    
-    contents:
-      reagents:
-      - ReagentId: chem.H2O
-        Quantity: 4
   - type: Sprite
     sprite: Objects/Drinks/thirteen_loko.rsi
   - type: Icon
