+++ Resources/Prototypes/Entities/mobs/human.yml
@@ -14,11 +14,14 @@
   - type: Hunger
   - type: Thirst
   # Organs
-  - type: Stomach
-    maxVolume: 100
-    digestionDelay: 20
+  - type: Solution
+    maxVol: 250
   - type: Bloodstream
-    maxVolume: 250
+    max_volume: 100
+  - type: Stomach
+    max_volume: 250
+    digestionDelay: 20
+
 
   - type: Inventory
   - type: Constructor
