+++ Resources/Prototypes/Reactions/drinks.yml
@@ -1,0 +1,103 @@
+- type: reaction
+  id: react.ManlyDorf
+  reactants:
+    chem.Beer:
+      amount: 1
+    chem.Ale:
+      amount: 2
+  products:
+    chem.ManlyDorf: 3
+
+- type: reaction
+  id: react.CubaLibre
+  reactants:
+    chem.Cola:
+      amount: 1
+    chem.Rum:
+      amount: 2
+  products:
+    chem.CubaLibre: 3
+
+- type: reaction
+  id: react.IrishCream
+  reactants:
+    chem.Cream:
+      amount: 1
+    chem.Whiskey:
+      amount: 2
+  products:
+    chem.IrishCream: 3
+
+- type: reaction
+  id: react.IrishCoffee
+  reactants:
+    chem.IrishCream:
+      amount: 2
+    chem.Coffee:
+      amount: 2
+  products:
+    chem.IrishCoffee: 4
+
+- type: reaction
+  id: react.IrishCarBomb
+  reactants:
+    chem.IrishCream:
+      amount: 1
+    chem.Ale:
+      amount: 1
+  products:
+    chem.IrishCarBomb: 2
+
+- type: reaction
+  id: react.B52
+  reactants:
+    chem.IrishCarBomb:
+      amount: 1
+    chem.Kahlua:
+      amount: 1
+    chem.Cognac:
+      amount: 1
+  products:
+    chem.B52: 3
+
+- type: reaction
+  id: react.AtomicBomb
+  reactants:
+    chem.B52:
+      amount: 10
+    chem.U:
+      amount: 1
+  products:
+    chem.AtomicBomb: 11
+
+- type: reaction
+  id: react.WhiskeyCola
+  reactants:
+    chem.Whiskey:
+      amount: 2
+    chem.Cola:
+      amount: 1
+  products:
+    chem.WhiskeyCola: 3
+
+- type: reaction
+  id: react.SyndicateBomb
+  reactants:
+    chem.WhiskeyCola:
+      amount: 1
+    chem.Beer:
+      amount: 1
+  products:
+    chem.SyndicateBomb: 2
+
+- type: reaction
+  id: react.Antifreeze
+  reactants:
+    chem.Vodka:
+      amount: 2
+    chem.Cream:
+      amount: 1
+    chem.Ice:
+      amount: 1
+  products:
+    chem.Antifreeze: 4
